requirements.txt
Flask == 3.1.2
icecream == 2.1.10
mysql-connector-python == 9.0.0
flask-session == 0.8.0